+++ src/shoutit/api/v3/CHANGELOG.md
@@ -16,6 +16,7 @@
 - Conversations (from type `public_chats`) can be now reported using `POST /misc/reports`
 - Allow attaching profiles in messages. Check [Message Attachment](https://github.com/shoutit/shoutit-api/wiki/Intro-to-Messaging#message-attachment)
 - Added `/conversation/{id}/media` to list media attachments (images and videos)
+- Added `/conversation/{id}/shouts` to list shout attachments
 
 
 ## Update (2016-05-10)